perf: Don’t fetch column for end of range
Previously when the range referred to was in a different file, the file would be loaded into a
buffer so that the exact column for the line ending a range could be provided in the linter message.
Removing this sped up one test by 10% and the loss is not really noticeable to the user.