feat: Add option to use ignore files like .gitignore
This uses the `ignore-walk` package to traverse a project’s paths, collecting all paths that are not
excluded by provided ignore files such as `.gitignore` or `.npmignore`, and passing the path list to
the `files` option of `jscpd`. This is only active when the linter scope is “project” and helps with
project-specific exclusions such as not comparing against builds or node_modules etc.